refactor(i18n): share one glossary override module across all five readers
The override layer was implemented twice: `applyOverrides` in
sync-glossary.ts and a `withOverrides` twin in .i18nrc.cjs, duplicated
because the lobe-i18n config is CommonJS and cannot load a TypeScript
module. Two copies of these rules is exactly the failure #1173 existed to
fix, where the translator's prompt enforced different terms than the
reviewer.

Move the implementation to scripts/i18n/glossary-overrides.cjs, the one
shape both sides can load: node `require`s it from the config, and the
TypeScript scripts import it through tsx. JSDoc types give the TS side
real types (`allowJs` is on), so `GlossaryOverrides` and both function
signatures still typecheck rather than degrading to `any`.

No behaviour change. Verified end to end:

  .i18nrc.cjs required from plain node CJS, effective/ moved aside so the
  raw-mirror fallback runs: 2804 pairs, Video and VIDEO retracted, no
  `- Video → null` rendered, Workflow retained
  pnpm i18n:glossary against the real ComfyUI_frontend locales: 2001
  effective pairs, es 198, every other locale 200 (tr 203), `Top` not
  promoted
  pnpm i18n:validate: no violations
  692 vitest passing; breaking the retraction branch in the single module
  fails both suites
  eslint clean, prettier clean on touched files, astro check 0 errors
